continuing custom iterators; bugs spotted
There is an issue when i place a value at the beginning of the list that it is given a larger address ( which remains in order of addresses dished out ) and causes any kind of address comparison to fail.  There is also another issue to note with begin() and end() returning the head and tail and I feel it should return one past or one before the first value to eliminate a != comparison stopping all logic when it hits the head.  Which is not the functionality im looking for, i would like for it to execute the logic on the last element as well.